Mirror BrAPI splitting of API into modules
Introduced a core package under the api level to mirror the BrAPI
specification's move to a system of related modules.
